\name{ryandexdirect-package}
\alias{ryandexdirect-package}
\alias{ryandexdirect}
\docType{package}
\title{
\packageTitle{ryandexdirect}
}
\description{
\packageDescription{ryandexdirect}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ryandexdirect}
\packageIndices{ryandexdirect}
}
\author{
\packageAuthor{ryandexdirect}

Maintainer: \packageMaintainer{ryandexdirect}
}

\keyword{ package }
